Fix external links reported by html-proofer
Many of these its a simple fix to use a new working URL but some are just not replaceable so I'm just removing for now. In some cases those things have long gone like FitnessClass (although can be seen a little through the 9mmedia.com site) and the PlayStation forums that was announced just a few months ago that it would shut down (URL doesn't work now as expected).