Use scrollIntoView to smoothly animate to top
This is backwards compatible so if a browser doesn't support scrollIntoView the default behavior with a named anchor will kick in. Otherwise this is used to override that. Safari does not support the smooth behavior (treats argument as boolean) but the default behavior of going to top is preserved. Chrome supports the smooth behavior so that is used there.