Replace only mixins with direct usage instead
I'm encountering a SASSC error when building the page but its hard to decipher what's causing it since it doesn't show the SASS line. It seems that updating dependencies fixes the issue and the build doesn't fail but since I started this, and we hardly use these mixins its not worth adding them as mixins and just use them directly. We only had 2 instances for the dark mode mixin so best to inline it as it saves us nothing. The media query breakpoints, we only use 2 of the 3 values so instead of a mixin and map, we'll just use 3 simple variables and reference directly. This minimizes SASS dependence on features that may or may not work in sass-embedded gem.